+++ backend/src/Logitar.Portal.Application/ApiKeys/Commands/UpdateApiKeyCommand.cs
@@ -66,7 +66,7 @@ internal class UpdateApiKeyCommandHandler : IRequestHandler<UpdateApiKeyCommand,
       }
     }
 
-    IEnumerable<FoundRole> roles = await _mediator.Send(new FindRolesQuery(apiKey.TenantId, payload.Roles, nameof(payload.Roles)), cancellationToken);
+    IReadOnlyCollection<FoundRole> roles = await _mediator.Send(new FindRolesQuery(apiKey.TenantId, payload.Roles, nameof(payload.Roles)), cancellationToken);
     foreach (FoundRole found in roles)
     {
       switch (found.Action)