+++ backend/src/Logitar.Portal.Application/Roles/Queries/FindRolesQuery.cs
@@ -5,7 +5,7 @@ using MediatR;
 
 namespace Logitar.Portal.Application.Roles.Queries;
 
-internal record FindRolesQuery : IRequest<IEnumerable<FoundRole>>
+internal record FindRolesQuery : IRequest<IReadOnlyCollection<FoundRole>>
 {
   public TenantId? TenantId { get; }
   public IEnumerable<RoleModification> Roles { get; }
@@ -24,7 +24,7 @@ internal record FindRolesQuery : IRequest<IEnumerable<FoundRole>>
   }
 }
 
-internal class FindRolesQueryHandler : IRequestHandler<FindRolesQuery, IEnumerable<FoundRole>>
+internal class FindRolesQueryHandler : IRequestHandler<FindRolesQuery, IReadOnlyCollection<FoundRole>>
 {
   private readonly IRoleRepository _roleRepository;
 
@@ -33,7 +33,7 @@ internal class FindRolesQueryHandler : IRequestHandler<FindRolesQuery, IEnumerab
     _roleRepository = roleRepository;
   }
 
-  public async Task<IEnumerable<FoundRole>> Handle(FindRolesQuery query, CancellationToken cancellationToken)
+  public async Task<IReadOnlyCollection<FoundRole>> Handle(FindRolesQuery query, CancellationToken cancellationToken)
   {
     int capacity = query.Roles.Count();
     Dictionary<RoleId, FoundRole> foundRoles = new(capacity);