+++ backend/src/Logitar.Portal.Domain/Senders/ISenderRepository.cs
@@ -6,8 +6,8 @@ public interface ISenderRepository
 {
   Task<Sender?> LoadAsync(Guid id, CancellationToken cancellationToken = default);
   Task<Sender?> LoadAsync(SenderId id, long? version, CancellationToken cancellationToken = default);
-  Task<IEnumerable<Sender>> LoadAsync(CancellationToken cancellationToken = default);
-  Task<IEnumerable<Sender>> LoadAsync(TenantId? tenantId, CancellationToken cancellationToken = default);
+  Task<IReadOnlyCollection<Sender>> LoadAsync(CancellationToken cancellationToken = default);
+  Task<IReadOnlyCollection<Sender>> LoadAsync(TenantId? tenantId, CancellationToken cancellationToken = default);
   Task<Sender?> LoadDefaultAsync(TenantId? tenantId, CancellationToken cancellationToken = default);
 
   Task SaveAsync(Sender sender, CancellationToken cancellationToken = default);