+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed Contract models and Domain aggregates.
 - Synchronized One-Time Password (OTP) namespaces.
 - Join Commands/Queries and their Handler.
+- Replace `IEnumerable<T>` by `IReadOnlyCollection<T>`.
 
 ### Removed